refactor(audio): restructure pipeline with cleaner naming and module boundaries
- move strip_fillers_and_repeats to clean_recording module
- rename process_recording -> process_whole_recording
- rename transcribe_chunk_samples -> process_chunk
- rename finalize_recording -> merge_all_chunks
- reorder functions: process_chunk and merge_all_chunks first
- add step comments to process_chunk continuing from merge_all_chunks
- move RecorderStartError to types.rs
