+++ src-tauri/src/audio/clean_recording.rs
@@ -61,6 +61,52 @@ fn strip_trailing_wake_word(text: &str, wake_word: &str) -> String {
     trimmed.to_string()
 }
 
+static FILLER_WORDS: &[&str] = &[
+    "euh", "hmm", "hm", "mmm", "uh", "um", "uhm", "umm", "uhh", "uhhh", "ah", "mm", "mh", "eh",
+    "ehh", "ha", "mm-hmm",
+];
+
+fn is_filler(word: &str) -> bool {
+    let normalized = word.trim_matches(|c: char| !c.is_alphanumeric());
+    if normalized.is_empty() {
+        return false;
+    }
+    let lower = normalized.to_lowercase();
+    FILLER_WORDS.contains(&lower.as_str())
+}
+
+pub(super) fn strip_fillers_and_repeats(text: &str) -> String {
+    let words: Vec<&str> = text.split_whitespace().filter(|w| !is_filler(w)).collect();
+    if words.is_empty() {
+        return String::new();
+    }
+
+    let mut result: Vec<&str> = Vec::with_capacity(words.len());
+    let mut i = 0;
+
+    while i < words.len() {
+        let current_lower = words[i].to_lowercase();
+        let mut count = 1;
+
+        while i + count < words.len() && words[i + count].to_lowercase() == current_lower {
+            count += 1;
+        }
+
+        if count >= 3 {
+            result.push(words[i]);
+            result.push(words[i + 1]);
+        } else {
+            for j in 0..count {
+                result.push(words[i + j]);
+            }
+        }
+
+        i += count;
+    }
+
+    result.join(" ")
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -162,6 +208,97 @@ mod tests {
         assert_eq!(
             strip_trailing_wake_word("bonjour alix validate.", "alix validate"),
             "bonjour"
+        );
+    }
+
+    #[test]
+    fn dedup_four_to_two() {
+        assert_eq!(strip_fillers_and_repeats("je je je je vais"), "je je vais");
+    }
+
+    #[test]
+    fn dedup_two_kept_unchanged() {
+        assert_eq!(strip_fillers_and_repeats("oui oui"), "oui oui");
+    }
+
+    #[test]
+    fn dedup_five_to_two() {
+        assert_eq!(
+            strip_fillers_and_repeats("the the the the the cat"),
+            "the the cat"
+        );
+    }
+
+    #[test]
+    fn dedup_three_to_two_case_insensitive() {
+        assert_eq!(
+            strip_fillers_and_repeats("Hello HELLO hello world"),
+            "Hello HELLO world"
+        );
+    }
+
+    #[test]
+    fn dedup_no_repetition() {
+        assert_eq!(
+            strip_fillers_and_repeats("normal sentence"),
+            "normal sentence"
+        );
+    }
+
+    #[test]
+    fn dedup_empty_string() {
+        assert_eq!(strip_fillers_and_repeats(""), "");
+    }
+
+    #[test]
+    fn dedup_single_word() {
+        assert_eq!(strip_fillers_and_repeats("word"), "word");
+    }
+
+    #[test]
+    fn dedup_multiple_groups() {
+        assert_eq!(
+            strip_fillers_and_repeats("the the the cat the the the dog"),
+            "the the cat the the dog"
+        );
+    }
+
+    #[test]
+    fn dedup_exactly_three_to_two() {
+        assert_eq!(
+            strip_fillers_and_repeats("hello hello hello world"),
+            "hello hello world"
+        );
+    }
+
+    #[test]
+    fn dedup_one_occurrence_unchanged() {
+        assert_eq!(strip_fillers_and_repeats("hello world"), "hello world");
+    }
+
+    #[test]
+    fn filler_isolated_removed() {
+        assert_eq!(strip_fillers_and_repeats("je euh vais"), "je vais");
+    }
+
+    #[test]
+    fn filler_repeated_fully_removed() {
+        assert_eq!(strip_fillers_and_repeats("euh euh euh bonjour"), "bonjour");
+    }
+
+    #[test]
+    fn filler_substring_in_real_word_kept() {
+        assert_eq!(
+            strip_fillers_and_repeats("aujourd'hui ah hammer"),
+            "aujourd'hui hammer"
+        );
+    }
+
+    #[test]
+    fn filler_mm_hmm_removed() {
+        assert_eq!(
+            strip_fillers_and_repeats("oui mm-hmm bonjour"),
+            "oui bonjour"
         );
     }
 }
